feat: Add tasks handler with CRUD operations and routing
